rename shared_handlers as docker_compose_utils
