+++ services/dynamic-sidecar/src/simcore_service_dynamic_sidecar/api/containers_extension.py
@@ -22,11 +22,11 @@ from servicelib.fastapi.requests_decorators import cancellable_request
 from servicelib.utils import logged_gather
 from simcore_sdk.node_ports_v2.port_utils import is_file_type
 
+from ..core.docker_compose_utils import write_file_and_run_command
 from ..core.docker_logs import start_log_fetching, stop_log_fetching
 from ..core.docker_utils import docker_client
 from ..core.rabbitmq import RabbitMQ
 from ..core.settings import DynamicSidecarSettings
-from ..core.shared_handlers import write_file_and_run_command
 from ..models.ports import PortTypeName
 from ..models.shared_store import SharedStore
 from ..modules import directory_watcher, nodeports
