+++ services/dynamic-sidecar/src/simcore_service_dynamic_sidecar/core/docker_compose_utils.py
